Fix message id race
It is possible that after sending a message, but before receiving
confirmation for it, remote party to sens a message, in which case lastmid
gets overwritten with mid of the new message. When confirmation for the
send message comes later on, lastmid is already different, leading to our
message being treated as a scrollback.

Fix that by keeping sent message id separately from received one.

Signed-off-by: Ivaylo Dimitrov <ivo.g.dimitrov.75@gmail.com>